% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushstroke.R
\name{brushstroke}
\alias{brushstroke}
\title{Draw a single brushstroke}
\usage{
brushstroke(x, y, width = 1, height = 0.5, angle = 0, xnoise = NULL,
  ynoise = NULL, nside = 10, col = "blue", curvature = NULL,
  border_col = NA, xstretch = 1, ystretch = 1)
}
\arguments{
\item{x}{x-coordinate of centre}

\item{y}{y-coordinate of centre}

\item{width}{width of rectangle}

\item{height}{height of rectangle}

\item{angle}{angle with positive x-axis}

\item{xnoise}{amount of noise to be added to vertical edges}

\item{ynoise}{amount of noise to be added to horizontal edges}

\item{nside}{number of points which can be adjusted by noise
on each of the four sides}

\item{col}{fill colour}

\item{curvature}{if positive, bends the middle of the rectangle
downwards. If negative, bends it upwards}

\item{border_col}{rectangle border colour (NA is transparent)}

\item{xstretch}{amount stretched in x direction}

\item{ystretch}{amount stretched in y direction; these are useful
when plotting with axes on different scales}
}
\value{
void
}
\description{
Draw a single brushstroke
}
\examples{
plot(1:10, 1:10, type="n", bty="n",xlab="",ylab="",yaxt="n", xaxt="n")
for (i in 1:10){
  brushstroke(i, 10, width=i/10, height=0.25, xnoise=0, ynoise=0, curvature=0)
  brushstroke(i, 9, width=0.5, height=i/10, xnoise=0, ynoise=0, curvature=0)
  brushstroke(i, 7, width=0.5, height=0.25, curvature=0, xnoise=(i-1)*0.05, ynoise=0)
  brushstroke(i, 6, width=0.5, height=0.25, curvature=0, xnoise=0, ynoise=0.01*(i-1))
  brushstroke(i, 5, width=0.5, height=0.25, curvature=0.1*(i-5), xnoise=0, ynoise=0)
  brushstroke(i, 3, width=0.5, height=0.25, curvature=0, xnoise=0, ynoise=0, angle=(i-5)/5 * pi/4)
  brushstroke(i, 2, width=0.5, height=0.25, curvature=0, xnoise=0, ynoise=0, col=rgb(0,0,1,i/10))
  brushstroke(i, 1, width=0.5, height=0.25)
}
}

