% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_plot_columns.R
\name{choose_plot_columns}
\alias{choose_plot_columns}
\title{Choose which columns of a data frame to plot}
\usage{
choose_plot_columns(X, default_angle = 0, default_alpha = 0.15, trace = T)
}
\arguments{
\item{X}{a data frame}

\item{default_angle}{the angle at which to plot the brushstrokes}

\item{default_alpha}{the alpha value (transparency) of the brushstroke colours}

\item{trace}{debugging flag}
}
\value{
a list with components x, y, angle, alpha, cols (colours),
names_x (name of x column),
names_y (name of y column)
}
\description{
This function chooses the columns to plot when a data frame
is passed to the koiplot() function.
}
\details{
This function takes a data frame as input. It chooses
\code{x} and \code{y} to be the first two numeric columns
encountered, starting from the left. It chooses the third numeric
column (if there is one) to be the angle variable. It chooses the
first non-numeric column to be the colour (after converting to a factor
and mapping with the rainbow() function) but if there is more than one
non-numeric column then it chooses the one with the smallest number
of distinct values among those which have at least 2 distinct values.

The function imputes NA values in \code{x} and \code{y} by the value
halfway between the max and min. It imputes NA values in the angle column
by the mean of the other angles. It does not do imputation in the colours.

If one or more of these columns consists entirely of NAs then the
function will fail.
}
\examples{
X <- data.frame(a=1:2, b=c(NA,4), c=factor(5:6), d=7:8)
choose_plot_columns(X)
X$b[2] <- NA
choose_plot_columns(X) # fails with warning message
}

