% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{gallery}
\alias{gallery}
\title{Gallery of extra examples}
\usage{
gallery(what = "")
}
\arguments{
\item{what}{a string naming the plot to be shown, or else all plots in the
gallery will be shown.}
}
\description{
A gallery of some examples of the use of koiplot(). May be
extended in future versions of the package.
}
\examples{
gallery()
}

