% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koiplotxy.R
\name{koiplotxy}
\alias{koiplotxy}
\title{Scatterplot using brushstrokes}
\usage{
koiplotxy(x, y, angle = NULL, alpha = 0.1, cols = "blue", width = NULL,
  height = NULL, curvature = NULL, rescale_angles = T, min_angle = -45,
  max_angle = 45, brushstroke_xnoise = NULL, brushstroke_ynoise = NULL,
  brushstroke_border_col = NA, nmax = 10000, add = F, ...)
}
\arguments{
\item{x}{vector of x-coordinates}

\item{y}{vector of y-coordinates}

\item{angle}{vector of angles for the brushstrokes (or a number)}

\item{alpha}{vector of alpha values for the brushstroke colours (or a number)}

\item{cols}{vector of colours for the brushstrokes (or a single colour)}

\item{width}{vector of widths for the brushstrokes (or a number)}

\item{height}{vector of heights for the brushstrokes (or a number)}

\item{min_angle}{minimum angle}

\item{max_angle}{maximum angle}

\item{brushstroke_xnoise}{vector of x-noise to be used for the brushstrokes}

\item{brushstroke_ynoise}{ditto y-noise}

\item{brushstroke_border_col}{ditto border colours}

\item{nmax}{maximum number of brushstrokes to be drawn (defaults to 10000 as more
becomes slow)}

\item{add}{if TRUE, don't draw a new plot (not recommended)}

\item{...}{extra arguments to plot() such as xlim, ylim, xaxt, title etc.}

\item{curavture}{ditto curvatures}
}
\description{
Makes a scatterplot of two specified vectors using brushstrokes from the
brushstroke() function instead of points
}
\details{
The arguments are passed to prepare_plot_data(). Then nmax random values from x #' and y are chosen
for plotting. Then a plot is made and the brushstrokes are added using brushstroke().
}
\examples{
koiplotxy(1:10, 1:10, cols=rep(c("red", "green"), 5),
angle=1:10, min_angle=0, max_angle=90,
brushstroke_border_col="blue",alpha=seq(0.1,1,0.1),
brushstroke_xnoise=0:9/10, brushstroke_ynoise=9:0/100,
curvature=0:9/10, xaxt="n", bty="n", col="orange",
main="Swatches")
# note that the "col" argument does nothing because plot() is not plotting any points
}

