% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare data for plotting}
\usage{
prepare_plot_data(x, y, angle = NULL, alpha = NULL, cols = "blue",
  width = NULL, height = NULL, curvature = NULL, rescale_angles = T,
  min_angle = -45, max_angle = 45, brushstroke_xnoise = NULL,
  brushstroke_ynoise = NULL, brushstroke_border_col = NA)
}
\arguments{
\item{x}{vector of x-coordinates}

\item{y}{vector of y-coordinates}

\item{angle}{vector of angles '}

\item{alpha}{vector of alpha values}

\item{cols}{vector of colours (character)}

\item{width}{vector of widths}

\item{height}{vector of heights}

\item{curvature}{vector of curvatures}

\item{rescale_angles}{boolean: if true, angles are to be rescaled to a specified range}

\item{min_angle}{min angle in range for angle rescaling (specified in degrees)}

\item{max_angle}{max angle in range for angle rescaling (specified in degrees)}

\item{brushstroke_xnoise}{amount of x-noise for brushstrokes (also a vector)}

\item{brushstroke_ynoise}{amount of y-noise for brushstrokes}

\item{brushstroke_border_col}{border colours for brushstrokes (NA = no borders plotted)}
}
\description{
Puts arguments into a form in which koiplotxy() can be applied
}
\details{
All arguments are recycled to the maximum length of x or y. If cols is numeric, a vector
of colours is created from it using rainbow().
}

